backend/requirements.txt
pydantic==2.12.5
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sentence-transformers==2.2.2
numpy==1.26.3
openai==1.12.0
pypdf
python-dotenv
